feat(routes): create regex for route paramenters extraction
