+++ my-server-nodejs/db.json
@@ -1,1 +1,1 @@
-{"users":[{"id":"098472ef-6fea-4ce9-9463-7fd66e90c6d2","name":"Kryvus","email":"Kryvus@test.com"}]}
+{"users":[{"id":"098472ef-6fea-4ce9-9463-7fd66e90c6d2","name":"Kryvus","email":"Kryvus@test.com"},{"id":"19775598-d29d-4c49-bc8f-71a3cb463c93","name":"Kryvus1","email":"Kryvus1@test.com"},{"id":"8c71062f-f22d-45b6-97f2-25c266a4e91e","name":"Kryvus2","email":"Kryvus2@test.com"}]}
