+++ my-server-nodejs/src/database.js
@@ -2,7 +2,7 @@ import fs from "node:fs/promises";
 
 const databasePath = new URL("../db.json", import.meta.url);
 
-console.log(databasePath);
+
 export class Database {
   #database = {};
