+++ my-server-nodejs/src/routes.js
@@ -1,0 +1,43 @@
+import { randomUUID } from "node:crypto";
+import { Database } from "./database.js";
+import { buildRoutePath } from "./utils/build-route-path.js";
+
+const database = new Database();
+
+export const routes = [
+  {
+    method: "GET",
+    path: buildRoutePath("/users"),
+    handler: (req, res) => {
+      const users = database.select("users");
+
+      return res.end(JSON.stringify(users));
+    },
+  },
+
+  {
+    method: "POST",
+    path: buildRoutePath("/users"),
+    handler: (req, res) => {
+      const { name, email } = req.body;
+
+      const user = {
+        id: randomUUID(),
+        name,
+        email,
+      };
+
+      database.insert("users", user);
+
+      return res.writeHead(201).end("201 Created");
+    },
+  },
+
+  {
+    method: 'DELETE',
+    path: buildRoutePath('/users/:id'),
+    handler: (req, res) => {
+        return res.end()
+    },
+  }
+];
