+++ my-server-nodejs/src/server.js
@@ -1,33 +1,20 @@
 import http from "node:http";
-import { randomUUID } from "node:crypto";
 import { json } from "./middlewares/json.js";
-import { Database } from "./database.js";
+import { routes } from "./routes.js";
 
-const database = new Database();
+
 
 const server = http.createServer(async (req, res) => {
   const { method, url } = req;
 
   await json(req, res);
 
-  if (method === "GET" && url === "/users") {
-    const users = database.select("users");
+  const route = routes.find(route => {
+    return route.method === method && route.path === url
+  })
 
-    return res.end(JSON.stringify(users));
-  }
-
-  if (method === "POST" && url === "/users") {
-    const { name, email } = req.body;
-
-    const user = {
-      id: randomUUID(),
-      name,
-      email,
-    };
-
-    database.insert("users", user);
-
-    return res.writeHead(201).end("201 Created");
+  if (route) {
+    return route.handler(req, res)
   }
 
   return res.writeHead(404).end();
