+++ my-server-nodejs/src/utils/build-route-path.js
@@ -1,0 +1,5 @@
+export function buildRoutePath(path) {
+    const routeParametersRegex = /:([a-zA-Z]+)/g
+
+    console.log(Array.from(path.matchAll(routeParametersRegex)))
+}
